requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
contourpy==1.3.1
croniter==6.0.0
cycler==0.12.1
discord-py-interactions==5.15.0
discord-py-slash-command==4.2.1
discord-typings==0.9.0
dnspython==2.7.0
emoji==2.14.1
fonttools==4.57.0
frozenlist==1.5.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
multidict==6.2.0
packaging==24.2
propcache==0.3.0
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.1
six==1.17.0
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.5.0
yarl==1.18.3
